requirements.txt
boto3==1.9.71
Flask==1.1.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.5
psutil==5.6.6
PyAudio==0.2.11
six==1.11.0
pyyaml=5.4.1
grpcio==1.34.0
pyusb==1.0.0b2
pvporcupine
iterators>=0.0.2
